Switch broker DLQ test to use a k8s Service
With upcoming changes to eventing DLQs will be defaulted to the
namespace of their parent object if the namespace isn't specified in the
ref. For some reason this causes a panic from the fake dynamic client
that doesn't occur if the namespace isn't specified.

It happens because ksvcs have been registered with the fake dynamic
client as a type, but I'm not really sure why they only panic when a
namespace is present.

xref https://github.com/knative/eventing/pull/5748 for the namespace
defaulting change
xref https://github.com/knative/eventing/issues/5185 for more discussion of
the changed behavior of client-go fake dynamic clients.