Add support for concurrent dispatch and configurable prefetch (#418)
* Reduce duplication in dispatcher test

* Add dispatcher prefetch annotation

* Extract brokertrigger test package

This lets us define different topologies and is a little more explicit
in the test what's being setup

* Dispatcher allows configuring prefetch, dispatches concurrently

Set a default of 10 and dispatch in parallel.

Had to change the unit tests to not care about ordering now that
messages are being handled asynchronously.

* Promote dispatcher fields to public

Easier to just construct with named fields rather than a
positional-argument constructor

* Use a worker pool for the dispatcher

This guarantees we don't exceed the concurrency factor set by the
prefetch count. This also allows us to wait for the workers to shut down
gracefully if we are cancelled.

* Inline prefetch size

* Use the same retry behavior when retrying to the broker