Update performance benchmarks post 1.0.0 release (#491)
* Update performance benchmarks post 1.0.0 release

I have confirmed that everything works as expected. I have removed all
"smart" scripting and went with the plain approach. I have also included
the results of the benchmarks in this repository, rather than
referencing files that were uploaded as a GitHub thread comment.

There are a few unexpected behaviours here which I have captured in the
README. Expanding on them below:

In previous iterations we didn't have any Receiver throughput values to
plot. I am not sure why, but I have confirmed that we now have them.

When plotting throughput, especially receiver throughput, gnuplot can
now take up to 5 minutes to load the graph. I suspect that this is due
to the solid area which gets rendered. We never had it before, so this
is a new & unexpected behaviour.

When publishing 1k messages per second, the receiver cannot keep up.
Increasing the CPU count from 1 to 2 doesn't help. I have attached a
screenshot in the PR which captures the expected performance. We may
want to troubleshoot, but for now we should just acknowledge that the 1k
msg/s performance test may work for short bursts, but the workload is
unsustainable (producers are constantly outpacing consumers).

Lastly, I have confirmed that
https://github.com/knative-sandbox/eventing-rabbitmq/issues/334#issuecomment-942411656
is spot-on from the broker perspective. Without it, the producer never
waits for the broker and simply keeps pushing messages, like UDP does
(fire & forget).

Signed-off-by: Gerhard Lazu <glazu@vmware.com>

* Add links to releases in benchmarking README

So that it's easier to find out whether we are running the latest ones.

Signed-off-by: Gerhard Lazu <glazu@vmware.com>