Add Makefile for running complicated tasks with a single command
For example, running end-to-end tests requires setting up dependencies,
a local KiND cluster, installing operators, waiting for components to be
ready, and then invoking `go test` with many flags. The entire flow is
captured in .github/workflows/kind-e2e.yaml, so how do we run it
locally? I think that there are some scripts which try to set everything
up, but they are not discoverable, pretty sure they diverged from the CI
workflow by now, and are not as composable as the Make equivalent.

So how does it work? Type make and follow along:

    build                       Build binaries with e2e tags
    dep-update                  Update any dependency
    e2e-setup                   Setup environment for e2e testing
    env                         Configure shell env - eval "$(make env)" OR rm .env && make .env && source .env
    k9s                         Terminal ncurses UI for K8s
    test-e2e-broker             Run Broker end-to-end tests
    test-e2e-publish            Run TestKoPublish end-to-end tests
    test-e2e-source             Run Source end-to-end tests
    test-unit                   Run unit tests

To run end-to-end tests, which is what I had to do in the context of
https://github.com/knative-sandbox/eventing-rabbitmq/pull/492 (this PR
is based on that one) I now only need to run the following:

    # This only needs to run once:
    make e2e-setup

    make test-e2e

    # I can run just a subset of tests with e.g.:
    make test-e2e-broker

This iteration:
- only support Linux & Darwin systems (does not support Windows)
- only supports x84_64 (does not support latest Macs with ARM silicon)
- defaults to the minimum supported K8s version (adding support for
  multiple K8s versions was out of scope)

This was only tested on Linux Ubuntu 20.04.3 LTS, and should be tested
on Darwin on Intel. Supporting Darwin on ARM should not be difficult,
but as I don't have access to one, I cannot test.

Signed-off-by: Gerhard Lazu <glazu@vmware.com>