Use RabbitmqCluster namespace in RabbitmqClusterReference (#492)
* Use RabbitmqCluster namespace in Exchange RabbitmqClusterReference

Fixes https://github.com/knative-sandbox/eventing-rabbitmq/issues/454

I will need to do the same with the Queue, but for now this is a good first step.

Use a temporary messaging-topology-operator version

The one that fixes the code-generator issue, as captured here:
https://github.com/rabbitmq/messaging-topology-operator/pull/257

Signed-off-by: Gerhard Lazu <glazu@vmware.com>

* Bump rabbitmq-operator & rabbitmq-messaging-topology-operator to latest

Same versions as used in go.mod

Signed-off-by: Gerhard Lazu <glazu@vmware.com>

* Fix rabbitmq-messaging-topology-conformance version

v1.2.2 is not out yet, use latest released version

Signed-off-by: Gerhard Lazu <glazu@vmware.com>

* Use RabbitmqCluster namespace in Queue RabbitmqClusterReference

Made the QueueLabels consistent with ExchangeLabels.

Signed-off-by: Gerhard Lazu <glazu@vmware.com>

* Use RabbitmqClusterReference namespace in Binding too

This is the last object, this should now work as requested:
https://github.com/knative-sandbox/eventing-rabbitmq/issues/454

Need to test this next, and possibly add an integration test - will ask
what the team consensus is on this.

As there is no messaging-topology-operator released with this fix, I am
pinning to the main branch version after this PR got merged:
https://github.com/rabbitmq/messaging-topology-operator/commit/5a753088044155e7e90c3c6ae64ee6b0d5d02136

Signed-off-by: Gerhard Lazu <glazu@vmware.com>

* Pass RabbitMQClusterName and RabbitMQClusterNamespace explicitly

Rather than grabbing them from Broker, and coupling Exchange, Queue &
Binding to Broker, we just pass the values in. Good suggestion @benmoss!

While Broker & Trigger are still used to build the Labels and the
OwnerReference, this is a step towards decoupling them completely.

Signed-off-by: Gerhard Lazu <glazu@vmware.com>

* Remove unused Broker from BindingArgs & QueueArgs

Great spot @benmoss, thank you!

Is there a linter that we have to spot these types of issues?

Signed-off-by: Gerhard Lazu <glazu@vmware.com>

* Revert BindingLabels & QueueLabels changes

There is already too much stuff happening in this PR. Thank you @benmoss
for suggesting that we split it in
https://github.com/knative-sandbox/eventing-rabbitmq/issues/513

Signed-off-by: Gerhard Lazu <glazu@vmware.com>

* Add unit tests for Exchange, Binding & Queue in RabbitMQ cluster namespace

Signed-off-by: Gerhard Lazu <glazu@vmware.com>

* Fix go.mod after main rebase

./hack/update-deps.sh FTW!

Signed-off-by: Gerhard Lazu <glazu@vmware.com>

* Add end-to-end test for Broker in different namespace to RMQ

Signed-off-by: Gerhard Lazu <glazu@vmware.com>