Remove ineffectual ctx assignment
This is a noop since we don't assign the context this function returns.
It likely was written here to reset the retry counter for retries to the
broker, but it also changes the strategy to exponential if the user was
using linear.

It's not really clear if we should be resetting the retry counter at
all, the semantics of this are vague, so for now let's just remove this.