Make install & test-e2e behaviour is more intuitive
install takes care of everything, and test-e2e also runs it - they are
self-contained targets.

The other test-e2e-* targets make assumptions that an install has run
before. The idea is that running a subset of e2e tests should be as fast
as possible, while running the entire e2e should be complete.

Signed-off-by: Gerhard Lazu <glazu@vmware.com>