Setup local host & run end-to-end tests with a single command (#525)
* Add Makefile for running complicated tasks with a single command

For example, running end-to-end tests requires setting up dependencies,
a local KiND cluster, installing operators, waiting for components to be
ready, and then invoking `go test` with many flags. The entire flow is
captured in .github/workflows/kind-e2e.yaml, so how do we run it
locally? I think that there are some scripts which try to set everything
up, but they are not discoverable, pretty sure they diverged from the CI
workflow by now, and are not as composable as the Make equivalent.

So how does it work? Type make and follow along:

    build                       Build binaries with e2e tags
    dep-update                  Update any dependency
    e2e-setup                   Setup environment for e2e testing
    env                         Configure shell env - eval "$(make env)" OR rm .env && make .env && source .env
    k9s                         Terminal ncurses UI for K8s
    test-e2e-broker             Run Broker end-to-end tests
    test-e2e-publish            Run TestKoPublish end-to-end tests
    test-e2e-source             Run Source end-to-end tests
    test-unit                   Run unit tests

To run end-to-end tests, which is what I had to do in the context of
https://github.com/knative-sandbox/eventing-rabbitmq/pull/492 (this PR
is based on that one) I now only need to run the following:

    # This only needs to run once:
    make e2e-setup

    make test-e2e

    # I can run just a subset of tests with e.g.:
    make test-e2e-broker

This iteration:
- only support Linux & Darwin systems (does not support Windows)
- only supports x84_64 (does not support latest Macs with ARM silicon)
- defaults to the minimum supported K8s version (adding support for
  multiple K8s versions was out of scope)

This was only tested on Linux Ubuntu 20.04.3 LTS, and should be tested
on Darwin on Intel. Supporting Darwin on ARM should not be difficult,
but as I don't have access to one, I cannot test.

Signed-off-by: Gerhard Lazu <glazu@vmware.com>

* Tell git to ingore .env and .config paths

We do not want to version-control these

Signed-off-by: Gerhard Lazu <glazu@vmware.com>

* Add KiND cluster config

Used in `make kind-cluster`

Signed-off-by: Gerhard Lazu <glazu@vmware.com>

* Install gh cli

This comes in handy for checking out PRs, working with Actions, etc.

Signed-off-by: Gerhard Lazu <glazu@vmware.com>

* Add descriptions for install-* targets

Allow versions to be changed just-in-time

Signed-off-by: Gerhard Lazu <glazu@vmware.com>

* Bump k9s to latest

Signed-off-by: Gerhard Lazu <glazu@vmware.com>

* Install kn CLI

After reading the first 6 chapters of Knative in Action, it became
obvious to me that this CLI is essential for working with Knative.

Signed-off-by: Gerhard Lazu <glazu@vmware.com>

* Move KO_URL after KO_BIN_DIR

On Windows on WSL with make 4.3 this fails. Not sure why it worked on
Linux with make 4.2.1 as it's clearly wrong. This fixes it.

Signed-off-by: Gerhard Lazu <glazu@vmware.com>

* Use make variables that expand only once

Variables declared with := get interpreted a single time while those
with = get evaluated every time they are invoked. We want just once
invocation.

Signed-off-by: Gerhard Lazu <glazu@vmware.com>

* Install serving part of the e2e setup

This is the easiest way to setup Sinks (they only need to be
addressable).

Signed-off-by: Gerhard Lazu <glazu@vmware.com>

* Install Kourier for ingress part of Knative Serving

Signed-off-by: Gerhard Lazu <glazu@vmware.com>

* Respect KUBECONFIG from env & do not hard-code to kind

Enables running against any K8s, helpful when testing against GKE.

Signed-off-by: Gerhard Lazu <glazu@vmware.com>

* Bump cert-manager to 1.5.4

This is the version that RabbitMQ messaging-topology-operator v1.2.1 was
tested against.

Related to https://github.com/knative-sandbox/eventing-rabbitmq/pull/370#discussion_r764031023

Signed-off-by: Gerhard Lazu <glazu@vmware.com>

* Fix missing OPEN

Great spot @benmoss!

Signed-off-by: Gerhard Lazu <glazu@vmware.com>

* Make test-e2e depend on KUBECONFIG, no more roundabouts

@benmoss made an excellent point in this comment, and I think that this
change address it.

https://github.com/knative-sandbox/eventing-rabbitmq/pull/525/files#r759457509

Signed-off-by: Gerhard Lazu <glazu@vmware.com>

* Remove kind setup scripts

These are now replaced by the Makefile, for context see
https://github.com/knative-sandbox/eventing-rabbitmq/pull/525

Signed-off-by: Gerhard Lazu <glazu@vmware.com>

* Use the existing context for the default install

Great suggestion @benmoss!

Signed-off-by: Gerhard Lazu <glazu@vmware.com>

* Make install & test-e2e behaviour is more intuitive

install takes care of everything, and test-e2e also runs it - they are
self-contained targets.

The other test-e2e-* targets make assumptions that an install has run
before. The idea is that running a subset of e2e tests should be as fast
as possible, while running the entire e2e should be complete.

Signed-off-by: Gerhard Lazu <glazu@vmware.com>

* Make the default help target more helpful for newcomers

That .env is always a gotcha... direnv doesn't make it better.

Signed-off-by: Gerhard Lazu <glazu@vmware.com>

* Rename build to test-compilation

I like it @benmoss!

https://github.com/knative-sandbox/eventing-rabbitmq/pull/525/files#r765968362

Signed-off-by: Gerhard Lazu <glazu@vmware.com>

* List test-unit-uncached in help

This is something that @ikvmw was mentioning yesterday. Making this
option visibile should help with WTF moments that originate in a stale cache.

Signed-off-by: Gerhard Lazu <glazu@vmware.com>

* Use Makefile in kind conformance standalone GitHub Action

The CI will continuously validate that everything is wired up correctly,
and devs using this locally will have the confidence that it just works.

Signed-off-by: Gerhard Lazu <glazu@vmware.com>

* Fix envsubst variables in kind.yaml

Signed-off-by: Gerhard Lazu <glazu@vmware.com>

* Do not link ko to gcloud

This is only necessary if pushing images to gcr.io, which is what *some*
of us do locally, but not everyone.

Signed-off-by: Gerhard Lazu <glazu@vmware.com>

* Uppercase all instances of KiND & K8s

Otherwise weird things happen, like KiND_CLUSTER_NAME vs
KIND_CLUSTER_NAME.

Signed-off-by: Gerhard Lazu <glazu@vmware.com>

* Export updated PATH so that all targets work correctly

Signed-off-by: Gerhard Lazu <glazu@vmware.com>

* Bump gh, k9s, kn & Knative versions to latest

Signed-off-by: Gerhard Lazu <glazu@vmware.com>

* Use make targets in kind e2e test GitHub Action

Signed-off-by: Gerhard Lazu <glazu@vmware.com>

* Install standalone broker for standalone conformance tests

Thanks @benmoss for spotting it!
https://github.com/knative-sandbox/eventing-rabbitmq/pull/525#issuecomment-994913634

Signed-off-by: Gerhard Lazu <glazu@vmware.com>

* Separate standalone conformance tests from the regular ones

Signed-off-by: Gerhard Lazu <glazu@vmware.com>

* Make test-conformance targets leaner

So that they are less likely to fail in GitHub Actions.

Signed-off-by: Gerhard Lazu <glazu@vmware.com>

* Convert standard conformance tests to use make targets

Signed-off-by: Gerhard Lazu <glazu@vmware.com>